[to be squashed with "Add the STORE module"] Rename store_local.h to store_locl.h
