+++ crypto/store/store_init.c
@@ -9,7 +9,7 @@
 
 #include <openssl/err.h>
 #include "internal/store.h"
-#include "store_local.h"
+#include "store_locl.h"
 
 static CRYPTO_ONCE store_init = CRYPTO_ONCE_STATIC_INIT;
 DEFINE_RUN_ONCE_STATIC(do_store_init)
