+++ crypto/store/store_locl.h
