+++ crypto/store/store_register.c
@@ -13,7 +13,7 @@
 
 #include <openssl/err.h>
 #include <openssl/lhash.h>
-#include "store_local.h"
+#include "store_locl.h"
 
 static CRYPTO_RWLOCK *registry_lock;
 static CRYPTO_ONCE registry_init = CRYPTO_ONCE_STATIC_INIT;
